requirements.txt
gymnasium
stable-baselines3[extra]
flappy-bird-gymnasium
tensorflow
torch-stubs